+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/bitbucketserver/BitbucketServerClient.kt
@@ -1,5 +1,6 @@
 package com.github.jensim.megamanipulator.actions.vcs.bitbucketserver
 
+import com.fasterxml.jackson.module.kotlin.readValue
 import com.github.jensim.megamanipulator.actions.localrepo.LocalRepoOperator
 import com.github.jensim.megamanipulator.actions.search.SearchResult
 import com.github.jensim.megamanipulator.actions.vcs.BitBucketPullRequestWrapper
@@ -7,13 +8,17 @@ import com.github.jensim.megamanipulator.actions.vcs.BitBucketRepoWrapping
 import com.github.jensim.megamanipulator.actions.vcs.PrActionStatus
 import com.github.jensim.megamanipulator.actions.vcs.PullRequestWrapper
 import com.github.jensim.megamanipulator.http.HttpClientProvider
+import com.github.jensim.megamanipulator.http.bodyAsText
+import com.github.jensim.megamanipulator.http.setBody
+import com.github.jensim.megamanipulator.http.unwrap
 import com.github.jensim.megamanipulator.project.lazyService
-import com.github.jensim.megamanipulator.settings.SerializationHolder
+import com.github.jensim.megamanipulator.settings.SerializationHolder.objectMapper
 import com.github.jensim.megamanipulator.settings.types.CloneType
 import com.github.jensim.megamanipulator.settings.types.CodeHostSettings.BitBucketSettings
 import com.intellij.openapi.project.Project
 import com.intellij.serviceContainer.NonInjectable
 import io.ktor.client.HttpClient
+import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.accept
 import io.ktor.client.request.delete
 import io.ktor.client.request.get
@@ -25,9 +30,6 @@ import io.ktor.client.statement.readText
 import io.ktor.http.ContentType
 import io.ktor.http.HttpStatusCode
 import io.ktor.http.contentType
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.decodeFromJsonElement
 import org.slf4j.LoggerFactory
 
 @SuppressWarnings("TooManyFunctions")
@@ -46,7 +48,6 @@ class BitbucketServerClient @NonInjectable constructor(
     private val httpClientProvider: HttpClientProvider by lazyService(project, httpClientProvider)
     private val localRepoOperator: LocalRepoOperator by lazyService(project, localRepoOperator)
 
-    private val json: Json = SerializationHolder.readableJson
     private val log = LoggerFactory.getLogger(javaClass)
 
     private suspend fun getDefaultReviewers(
@@ -54,7 +55,10 @@ class BitbucketServerClient @NonInjectable constructor(
         settings: BitBucketSettings,
         pullRequest: BitBucketPullRequestWrapper
     ): List<BitBucketUser> {
-        return client.get("${settings.baseUrl}/rest/default-reviewers/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/reviewers?sourceRepoId=${pullRequest.bitbucketPR.fromRef.repository.id}&targetRepoId=${pullRequest.bitbucketPR.toRef.repository.id}&sourceRefId=${pullRequest.bitbucketPR.fromRef.id}&targetRefId=${pullRequest.bitbucketPR.toRef.id}")
+        val url =
+            "${settings.baseUrl}/rest/default-reviewers/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/reviewers?sourceRepoId=${pullRequest.bitbucketPR.fromRef.repository.id}&targetRepoId=${pullRequest.bitbucketPR.toRef.repository.id}&sourceRefId=${pullRequest.bitbucketPR.fromRef.id}&targetRefId=${pullRequest.bitbucketPR.toRef.id}"
+        val response: HttpResponse = client.get(url)
+        return response.unwrap()
     }
 
     suspend fun addDefaultReviewers(
@@ -82,19 +86,20 @@ class BitbucketServerClient @NonInjectable constructor(
         val bitBucketSourceRepo = if (repo == sourceRepo) bitBucketRepo else {
             getRepo(client, settings, sourceRepo)
         }
-        return client.get("${settings.baseUrl}/rest/default-reviewers/1.0/projects/${repo.project}/repos/${repo.repo}/reviewers?sourceRepoId=${bitBucketSourceRepo.id}&targetRepoId=${bitBucketRepo.id}&sourceRefId=$fromBranchRef&targetRefId=$toBranchRef") {
+        val url = "${settings.baseUrl}/rest/default-reviewers/1.0/projects/${repo.project}/repos/${repo.repo}/reviewers?sourceRepoId=${bitBucketSourceRepo.id}&targetRepoId=${bitBucketRepo.id}&sourceRefId=$fromBranchRef&targetRefId=$toBranchRef"
+        val response: HttpResponse = client.get(url) {
             accept(ContentType.Application.Json)
         }
+        return response.unwrap()
     }
 
     suspend fun getRepo(searchResult: SearchResult, settings: BitBucketSettings): BitBucketRepoWrapping {
-        val client: HttpClient =
-            httpClientProvider.getClient(searchResult.searchHostName, searchResult.codeHostName, settings)
+        val client: HttpClient = httpClientProvider.getClient(searchResult.searchHostName, searchResult.codeHostName, settings)
         val repo = getRepo(client, settings, searchResult)
-        val defaultBranch =
-            client.get<BitBucketDefaultBranch>("${settings.baseUrl}/rest/api/1.0/projects/${repo.project?.key!!}/repos/${repo.slug}/default-branch") {
-                accept(ContentType.Application.Json)
-            }.displayId
+        val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/projects/${repo.project?.key!!}/repos/${repo.slug}/default-branch") {
+            accept(ContentType.Application.Json)
+        }
+        val defaultBranch = response.unwrap<BitBucketDefaultBranch>().displayId
         return BitBucketRepoWrapping(
             searchHost = searchResult.searchHostName,
             codeHost = searchResult.codeHostName,
@@ -104,9 +109,10 @@ class BitbucketServerClient @NonInjectable constructor(
     }
 
     private suspend fun getRepo(client: HttpClient, settings: BitBucketSettings, repo: SearchResult): BitBucketRepo {
-        return client.get("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
+        val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
             accept(ContentType.Application.Json)
         }
+        return response.unwrap()
     }
 
     suspend fun createPr(
@@ -116,8 +122,9 @@ class BitbucketServerClient @NonInjectable constructor(
         repo: SearchResult
     ): PullRequestWrapper {
         val client: HttpClient = httpClientProvider.getClient(repo.searchHostName, repo.codeHostName, settings)
-        val defaultBranch =
-            client.get<BitBucketDefaultBranch>("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}/default-branch").id
+        val url = "${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}/default-branch"
+        val response: HttpResponse = client.get(url)
+        val defaultBranch: String = response.unwrap<BitBucketDefaultBranch>().id
         val localBranch: String = localRepoOperator.getBranch(repo)!!
         val fork: Pair<String, String>? = localRepoOperator.getForkProject(repo)
         val fromProject = fork?.first ?: repo.project
@@ -151,17 +158,16 @@ class BitbucketServerClient @NonInjectable constructor(
             ),
             reviewers = reviewers,
         )
-        val prResponse: HttpResponse =
-            client.post("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}/pull-requests") {
-                contentType(ContentType.Application.Json)
-                accept(ContentType.Application.Json)
-                body = request
-            }
+        val prResponse: HttpResponse = client.post("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}/pull-requests") {
+            contentType(ContentType.Application.Json)
+            accept(ContentType.Application.Json)
+            setBody(request)
+        }
         if (prResponse.status.value >= 300) {
             log.error("Not OK response creating Pull Request")
         }
-        val prRaw = prResponse.readText()
-        val pr: BitBucketPullRequest = json.decodeFromString(BitBucketPullRequest.serializer(), prRaw)
+        val prRaw: String = prResponse.bodyAsText()
+        val pr: BitBucketPullRequest = objectMapper.readValue(prRaw)
 
         return BitBucketPullRequestWrapper(
             searchHost = repo.searchHostName,
@@ -190,10 +196,10 @@ class BitbucketServerClient @NonInjectable constructor(
         val response: HttpResponse = client.put("${settings.baseUrl}/rest/api/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/pull-requests/${pullRequest.bitbucketPR.id}") {
             contentType(ContentType.Application.Json)
             accept(ContentType.Application.Json)
-            body = pullRequest.bitbucketPR.copy(author = null)
+            setBody(pullRequest.bitbucketPR.copy(author = null))
         }
         return if (response.status.value >= 300) {
-            PrActionStatus(success = false, msg = "Failed updating PR due to: '${response.readText()}'")
+            PrActionStatus(success = false, msg = "Failed updating PR due to: '${response.bodyAsText()}'")
         } else {
             PrActionStatus(success = true)
         }
@@ -215,21 +221,22 @@ class BitbucketServerClient @NonInjectable constructor(
         val collector = ArrayList<PullRequestWrapper>()
         var start = 0L
         while (collector.size < limit) {
-            val response: BitBucketPage = try {
-                client.get("${settings.baseUrl}/rest/api/1.0/dashboard/pull-requests?start=$start&limit=100$state$role") {
+            val page: BitBucketPage = try {
+                val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/dashboard/pull-requests?start=$start&limit=100$state$role") {
                     accept(ContentType.Application.Json)
                 }
+                response.unwrap()
             } catch (e: Exception) {
                 log.error("Failed fetching pull requests from bitbucket", e)
                 break
             }
-            response.message?.let {
+            page.message?.let {
                 println("Message received $it")
             }
             collector.addAll(
-                response.values.orEmpty().map { raw ->
-                    val pr = json.decodeFromJsonElement<BitBucketPullRequest>(raw)
-                    val prString = json.encodeToString(raw)
+                page.values.orEmpty().map { raw ->
+                    val pr = objectMapper.treeToValue(raw, BitBucketPullRequest::class.java)
+                    val prString = objectMapper.writeValueAsString(raw)
                     BitBucketPullRequestWrapper(
                         searchHost = searchHostName,
                         codeHost = codeHostName,
@@ -238,10 +245,10 @@ class BitbucketServerClient @NonInjectable constructor(
                     )
                 }
             )
-            if (response.isLastPage != false) {
+            if (page.isLastPage != false) {
                 break
             } else {
-                start += response.size ?: 0
+                start += page.size ?: 0
             }
         }
         return collector
@@ -259,21 +266,21 @@ class BitbucketServerClient @NonInjectable constructor(
         try {
             val urlString =
                 "${settings.baseUrl}/rest/api/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/pull-requests/${pullRequest.bitbucketPR.id}/decline?version=${pullRequest.bitbucketPR.version}"
-            val response = client.post<HttpResponse>(urlString) {
+            val response: HttpResponse = client.post(urlString) {
                 contentType(ContentType.Application.Json)
-                body = emptyMap<String, String>()
+                setBody(emptyMap<String, String>())
             }
             if (response.status.value >= 300) {
-                log.error("Failed declining PullRequest ${response.readText()}")
+                log.error("Failed declining PullRequest ${response.bodyAsText()}")
                 return PrActionStatus(false)
             }
             if (dropFork && pullRequest.isFork()) {
                 val repository = pullRequest.bitbucketPR.fromRef.repository
                 // Get open PRs
-                val page: BitBucketPage =
-                    client.get("${settings.baseUrl}/rest/api/1.0/projects/${repository.project?.key!!}/repos/${repository.slug}/pull-requests?direction=OUTGOING&state=OPEN") {
-                        accept(ContentType.Application.Json)
-                    }
+                val pageResponse: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/projects/${repository.project?.key!!}/repos/${repository.slug}/pull-requests?direction=OUTGOING&state=OPEN") {
+                    accept(ContentType.Application.Json)
+                }
+                val page: BitBucketPage = pageResponse.unwrap()
                 if ((page.size ?: 0) == 0) {
                     return deletePrivateRepo(
                         BitBucketRepoWrapping(
@@ -305,10 +312,10 @@ class BitbucketServerClient @NonInjectable constructor(
         // https://docs.atlassian.com/bitbucket-server/rest/5.8.0/bitbucket-branch-rest.html#idm45555984542992
         val response: HttpResponse = client.delete("${settings.baseUrl}/rest/branch-utils/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/branches") {
             contentType(ContentType.Application.Json)
-            body = BitBucketRemoveBranchRequest(name = pullRequest.bitbucketPR.fromRef.id, dryRun = false)
+            setBody(BitBucketRemoveBranchRequest(name = pullRequest.bitbucketPR.fromRef.id, dryRun = false))
         }
         return if (response.status.value >= 300) {
-            PrActionStatus(false, response.readText())
+            PrActionStatus(false, response.bodyAsText())
         } else {
             PrActionStatus(true)
         }
@@ -317,31 +324,30 @@ class BitbucketServerClient @NonInjectable constructor(
     suspend fun createFork(settings: BitBucketSettings, repo: SearchResult): String? {
         val client: HttpClient = httpClientProvider.getClient(repo.searchHostName, repo.codeHostName, settings)
         val bbRepo: BitBucketRepo = try {
-            if (repo.project.toLowerCase() == "~${settings.username.toLowerCase()}") {
+            if (repo.project.lowercase() == "~${settings.username.lowercase()}") {
                 // is private repo
-                client.get("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
+                val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
                     accept(ContentType.Application.Json)
                 }
+                response.unwrap()
             } else {
                 // If repo with prefix already exists..
-                client.get("${settings.baseUrl}/rest/api/1.0/users/${settings.username}/repos/${repo.repo}") {
+                val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/users/${settings.username}/repos/${repo.repo}") {
                     accept(ContentType.Application.Json)
                 }
+                response.unwrap()
             }
         } catch (e: Exception) {
             log.warn("Failed finding fork", e)
             // Repo does not exist - lets fork
             null
         } ?: try {
-            client.post("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
+            val response: HttpResponse = client.post("${settings.baseUrl}/rest/api/1.0/projects/${repo.project}/repos/${repo.repo}") {
                 contentType(ContentType.Application.Json)
                 accept(ContentType.Application.Json)
-                body = mapOf<String, String>()
-//                body = BitBucketForkRequest(
-//                        slug = repo.repo,
-//                        project = BitBucketProjectRequest(key = "~${settings.username}")
-//                )
+                setBody(mapOf<String, String>())
             }
+            response.unwrap()
         } catch (e: Throwable) {
             log.error("Failed forking repo", e)
             throw e
@@ -363,11 +369,12 @@ class BitbucketServerClient @NonInjectable constructor(
         val collector = HashSet<BitBucketRepo>()
         while (true) {
             val urlString = "${settings.baseUrl}/rest/api/1.0/users/${settings.username}/repos?start=$start"
-            val page: BitBucketPage = client.get(urlString) {
+            val response: HttpResponse = client.get(urlString) {
                 accept(ContentType.Application.Json)
             }
+            val page: BitBucketPage = response.unwrap()
             page.values.orEmpty()
-                .map { json.decodeFromJsonElement<BitBucketRepo>(it) }
+                .map { objectMapper.treeToValue(it, BitBucketRepo::class.java) }
                 .filter { it.origin != null }
                 .forEach { collector.add(it) }
             if (page.isLastPage != false) break
@@ -375,10 +382,10 @@ class BitbucketServerClient @NonInjectable constructor(
         }
         return collector.filter {
             // TODO FILTER if repo head or branch commits is ahead of origin
-            val page: BitBucketPage =
-                client.get("${settings.baseUrl}/rest/api/1.0/projects/${it.project?.key}/repos/${it.slug}/pull-requests?direction=OUTGOING&state=OPEN") {
-                    accept(ContentType.Application.Json)
-                }
+            val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/projects/${it.project?.key}/repos/${it.slug}/pull-requests?direction=OUTGOING&state=OPEN") {
+                accept(ContentType.Application.Json)
+            }
+            val page: BitBucketPage = response.unwrap()
             page.size == 0
         }.map {
             BitBucketRepoWrapping(
@@ -396,14 +403,14 @@ class BitbucketServerClient @NonInjectable constructor(
      */
     suspend fun deletePrivateRepo(repo: BitBucketRepoWrapping, settings: BitBucketSettings): PrActionStatus {
         val client: HttpClient = httpClientProvider.getClient(repo.getSearchHost(), repo.getCodeHost(), settings)
-        val response =
-            client.delete<HttpResponse>("${settings.baseUrl}/rest/api/1.0/users/${settings.username}/repos/${repo.getRepo()}") {
+        val response: HttpResponse =
+            client.delete("${settings.baseUrl}/rest/api/1.0/users/${settings.username}/repos/${repo.getRepo()}") {
                 contentType(ContentType.Application.Json)
                 accept(ContentType.Application.Json)
-                body = emptyMap<String, String>()
+                setBody(emptyMap<String, String>())
             }
         return if (response.status.value >= 300) {
-            val msg = "Failed deleting private repo ${repo.asPathString()} due to: '${response.readText()}'"
+            val msg = "Failed deleting private repo ${repo.asPathString()} due to: '${response.bodyAsText()}'"
             log.error(msg)
             PrActionStatus(success = false, msg = msg)
         } else {
@@ -415,20 +422,19 @@ class BitbucketServerClient @NonInjectable constructor(
         // https://docs.atlassian.com/bitbucket-server/rest/7.10.0/bitbucket-rest.html#idp323
         val client: HttpClient =
             httpClientProvider.getClient(pullRequest.searchHostName(), pullRequest.codeHostName(), settings)
-        val response =
-            client.post<HttpResponse>("${settings.baseUrl}/rest/api/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/pull-requests/${pullRequest.bitbucketPR.id}/comments") {
-                contentType(ContentType.Application.Json)
-                accept(ContentType.Application.Json)
-                body = BitBucketComment(text = comment)
-            }
+        val response: HttpResponse = client.post("${settings.baseUrl}/rest/api/1.0/projects/${pullRequest.project()}/repos/${pullRequest.baseRepo()}/pull-requests/${pullRequest.bitbucketPR.id}/comments") {
+            contentType(ContentType.Application.Json)
+            accept(ContentType.Application.Json)
+            setBody(BitBucketComment(text = comment))
+        }
         if (response.status.value >= 300) {
-            log.error("Failed commenting due to: '${response.readText()}'")
+            log.error("Failed commenting due to: '${response.bodyAsText()}'")
         }
     }
 
     suspend fun validateAccess(searchHost: String, codeHost: String, settings: BitBucketSettings): String = try {
         val client: HttpClient = httpClientProvider.getClient(searchHost, codeHost, settings)
-        val response = client.get<HttpResponse>("${settings.baseUrl}/rest/api/1.0/inbox/pull-requests/count") {
+        val response: HttpResponse = client.get("${settings.baseUrl}/rest/api/1.0/inbox/pull-requests/count") {
             accept(ContentType.Application.Json)
         }
         val desc = HttpStatusCode.fromValue(response.status.value).description
@@ -462,14 +468,16 @@ class BitbucketServerClient @NonInjectable constructor(
         val response: HttpResponse = client.put("${settings.baseUrl}/rest/api/1.0/projects/$projectKey/repos/$repositorySlug/pull-requests/$pullRequestId/participants/$userSlug") {
             contentType(ContentType.Application.Json)
             accept(ContentType.Application.Json)
-            body = BitBucketParticipantStatusRequest(
-                user = BitBucketUser(name = settings.username),
-                approved = status == BitBucketPullRequestStatus.APPROVED,
-                status = status
+            setBody(
+                BitBucketParticipantStatusRequest(
+                    user = BitBucketUser(name = settings.username),
+                    approved = status == BitBucketPullRequestStatus.APPROVED,
+                    status = status
+                )
             )
         }
         return if (response.status.value >= 300) {
-            PrActionStatus(success = false, msg = "Failed setting pr status due to: '${response.readText()}'")
+            PrActionStatus(success = false, msg = "Failed setting pr status due to: '${response.bodyAsText()}'")
         } else {
             PrActionStatus(success = true)
         }
@@ -489,7 +497,7 @@ class BitbucketServerClient @NonInjectable constructor(
             header("X-Atlassian-Token", "no-check")
         }
         return if (response.status.value >= 300) {
-            PrActionStatus(success = false, msg = "Failed merging due to: '${response.readText()}'")
+            PrActionStatus(success = false, msg = "Failed merging due to: '${response.bodyAsText()}'")
         } else {
             PrActionStatus(success = true)
         }