+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/gitlab/GitLabProject.kt
@@ -1,8 +1,7 @@
 package com.github.jensim.megamanipulator.actions.vcs.gitlab
 
-import kotlinx.serialization.Serializable
 
-@Serializable
+
 @Suppress("ConstructorParameterNaming")
 data class GitLabProject(
     val id: Long,