+++ src/main/kotlin/com/github/jensim/megamanipulator/http/HttpClientProvider.kt
@@ -1,5 +1,6 @@
 package com.github.jensim.megamanipulator.http
 
+import com.fasterxml.jackson.jr.ob.JSON
 import com.github.jensim.megamanipulator.actions.NotificationsOperator
 import com.github.jensim.megamanipulator.project.lazyService
 import com.github.jensim.megamanipulator.settings.SerializationHolder
@@ -18,23 +19,27 @@ import com.intellij.openapi.project.Project
 import com.intellij.serviceContainer.NonInjectable
 import io.ktor.client.HttpClient
 import io.ktor.client.HttpClientConfig
-import io.ktor.client.engine.apache.Apache
-import io.ktor.client.engine.apache.ApacheEngineConfig
+import io.ktor.client.call.receive
+import io.ktor.client.engine.cio.CIO
+import io.ktor.client.engine.cio.CIOEngineConfig
 import io.ktor.client.features.HttpTimeout
 import io.ktor.client.features.defaultRequest
+import io.ktor.client.features.json.JacksonSerializer
 import io.ktor.client.features.json.JsonFeature
-import io.ktor.client.features.json.serializer.KotlinxSerializer
 import io.ktor.client.features.logging.DEFAULT
 import io.ktor.client.features.logging.LogLevel
 import io.ktor.client.features.logging.Logger
 import io.ktor.client.features.logging.Logging
+import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.header
 import io.ktor.client.request.headers
-import org.apache.http.conn.ssl.NoopHostnameVerifier
-import org.apache.http.conn.ssl.TrustSelfSignedStrategy
+import io.ktor.client.statement.HttpResponse
+import io.ktor.client.statement.readText
+import io.ktor.client.statement.request
+import io.ktor.http.isSuccess
 import org.apache.http.conn.ssl.TrustStrategy
-import org.apache.http.ssl.SSLContextBuilder
 import java.security.cert.X509Certificate
+import javax.net.ssl.X509TrustManager
 
 class HttpClientProvider @NonInjectable constructor(
     project: Project,
@@ -58,43 +63,30 @@ class HttpClientProvider @NonInjectable constructor(
         override fun isTrusted(p0: Array<out X509Certificate>?, p1: String?): Boolean = true
     }
 
-    private fun bakeClient(installs: HttpClientConfig<ApacheEngineConfig>.() -> Unit): HttpClient = HttpClient(Apache) {
-        install(JsonFeature) {
-            serializer = KotlinxSerializer(
-                json = SerializationHolder.compactJson
-            )
+    private fun bakeClient(installs: HttpClientConfig<CIOEngineConfig>.() -> Unit): HttpClient = HttpClient(CIO) {
+        /*
+        install(ContentNegotiation) {
+            jackson {}
+        }
+         */
+        install(JsonFeature){
+            this.serializer = JacksonSerializer(jackson = SerializationHolder.objectMapper)
         }
         install(Logging) {
             logger = Logger.DEFAULT
-            level = LogLevel.HEADERS
+            level = LogLevel.ALL
         }
         installs()
     }
 
-    private fun HttpClientConfig<ApacheEngineConfig>.trustSelfSignedClient() {
+    private fun HttpClientConfig<CIOEngineConfig>.trustAnyClient() {
         engine {
-            customizeClient {
-                setSSLContext(
-                    SSLContextBuilder
-                        .create()
-                        .loadTrustMaterial(TrustSelfSignedStrategy())
-                        .build()
-                )
-                setSSLHostnameVerifier(NoopHostnameVerifier())
-            }
-        }
-    }
-
-    private fun HttpClientConfig<ApacheEngineConfig>.trustAnyClient() {
-        engine {
-            customizeClient {
-                setSSLContext(
-                    SSLContextBuilder
-                        .create()
-                        .loadTrustMaterial(TrustAnythingStrategy())
-                        .build()
-                )
-                setSSLHostnameVerifier(NoopHostnameVerifier())
+            https {
+                trustManager = object : X509TrustManager {
+                    override fun checkClientTrusted(p0: Array<out X509Certificate>?, p1: String?) = Unit
+                    override fun checkServerTrusted(p0: Array<out X509Certificate>?, p1: String?) = Unit
+                    override fun getAcceptedIssuers(): Array<X509Certificate>? = null
+                }
             }
         }
     }
@@ -129,15 +121,14 @@ class HttpClientProvider @NonInjectable constructor(
 
     fun getClient(httpsOverride: HttpsOverride?, auth: HostWithAuth, password: String): HttpClient {
         return bakeClient {
-
             install(HttpTimeout) {
                 connectTimeoutMillis = 1_000
                 requestTimeoutMillis = 60_000
                 socketTimeoutMillis = 60_000
             }
             when (httpsOverride) {
-                HttpsOverride.ALLOW_SELF_SIGNED_CERT -> trustSelfSignedClient()
                 HttpsOverride.ALLOW_ANYTHING -> trustAnyClient()
+                else -> {}
             }
             auth.getAuthHeaderValue(password)?.let {
                 installBasicAuth(it)
@@ -145,11 +136,27 @@ class HttpClientProvider @NonInjectable constructor(
         }
     }
 
-    private fun HttpClientConfig<ApacheEngineConfig>.installBasicAuth(headerValue: String) {
+    private fun HttpClientConfig<CIOEngineConfig>.installBasicAuth(headerValue: String) {
         defaultRequest {
             headers {
                 header("Authorization", headerValue)
             }
         }
     }
+}
+
+suspend inline fun <reified T> HttpResponse.unwrap(): T {
+    if (status.isSuccess()) {
+        // return body()
+        return receive()
+    } else {
+        val body: String = bodyAsText()
+        throw RuntimeException("Respone status ${status.value} from ${request.url} with message: $body")
+    }
+}
+
+
+suspend fun HttpResponse.bodyAsText() = this.readText()
+inline fun <reified T : Any> HttpRequestBuilder.setBody(t: T) {
+    this.body = t
 }