+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/SerializationHolder.kt
@@ -1,24 +1,23 @@
 package com.github.jensim.megamanipulator.settings
 
-import kotlinx.serialization.json.Json
+import com.fasterxml.jackson.databind.DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT
+import com.fasterxml.jackson.databind.DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT
+import com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.SerializationFeature.INDENT_OUTPUT
+import com.fasterxml.jackson.module.kotlin.KotlinModule
 
 object SerializationHolder {
 
-    val compactJson: Json by lazy {
-        Json {
-            isLenient = true
-            ignoreUnknownKeys = true
-            prettyPrint = false
-            encodeDefaults = false
-        }
+    fun new() = ObjectMapper().apply {
+        configure(FAIL_ON_UNKNOWN_PROPERTIES, false)
+        configure(ACCEPT_EMPTY_STRING_AS_NULL_OBJECT, true)
+        configure(ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT, true)
+        registerModule(KotlinModule.Builder().build())
     }
-    val readableJson: Json by lazy {
-        Json {
-            isLenient = true
-            ignoreUnknownKeys = true
-            prettyPrint = true
-            prettyPrintIndent = "  "
-            encodeDefaults = true
-        }
+
+    val objectMapper = new()
+    val readable = new().apply {
+        configure(INDENT_OUTPUT, true)
     }
 }