+++ src/test/kotlin/com/github/jensim/megamanipulator/settings/SettingsFileOperatorTest.kt
@@ -1,13 +1,20 @@
 package com.github.jensim.megamanipulator.settings
 
-import com.github.jensim.megamanipulator.settings.SerializationHolder.readableJson
+import com.fasterxml.jackson.databind.JsonNode
 import com.github.jensim.megamanipulator.settings.types.CodeHostSettings
 import com.github.jensim.megamanipulator.settings.types.MegaManipulatorSettings
 import com.github.jensim.megamanipulator.settings.types.SearchHostSettings
-import com.github.ricky12awesome.jss.encodeToSchema
-import com.github.ricky12awesome.jss.globalJson
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
+import com.github.victools.jsonschema.generator.OptionPreset
+import com.github.victools.jsonschema.generator.SchemaGenerator
+import com.github.victools.jsonschema.generator.SchemaGeneratorConfig
+import com.github.victools.jsonschema.generator.SchemaGeneratorConfigBuilder
+import com.github.victools.jsonschema.generator.SchemaVersion
+import com.github.victools.jsonschema.module.jackson.JacksonModule
+import com.github.victools.jsonschema.module.jakarta.validation.JakartaValidationModule
+import com.github.victools.jsonschema.module.jakarta.validation.JakartaValidationOption.INCLUDE_PATTERN_EXPRESSIONS
+import com.github.victools.jsonschema.module.jakarta.validation.JakartaValidationOption.NOT_NULLABLE_FIELD_IS_REQUIRED
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Disabled
@@ -46,8 +53,8 @@ class SettingsFileOperatorTest {
         // given
 
         // when
-        val json = readableJson.encodeToString(testData)
-        val deserialized: MegaManipulatorSettings = readableJson.decodeFromString(json)
+        val json = SerializationHolder.readable.writeValueAsString(testData)
+        val deserialized: MegaManipulatorSettings = SerializationHolder.readable.readValue(json, MegaManipulatorSettings::class.java)
 
         // then
         assertEquals(deserialized, testData)
@@ -74,7 +81,7 @@ class SettingsFileOperatorTest {
     fun `test default settings`() {
         val defaultFileContent = File("src/main/resources/base-files/soft/mega-manipulator.json").readText()
 
-        val fromFile = readableJson.decodeFromString<MegaManipulatorSettings>(defaultFileContent)
+        val fromFile = SerializationHolder.readable.readValue(defaultFileContent, MegaManipulatorSettings::class.java)
         assertEquals(fromFile, testData)
     }
 
@@ -83,9 +90,15 @@ class SettingsFileOperatorTest {
     fun `generate json schema and compare to file`() {
         val baseFile = File("src/main/resources/base-files/hard/mega-manipulator-schema.json")
         val fileSystemSchema: String = baseFile.readText().trim()
-        val generatedSchema: String = globalJson.encodeToSchema(MegaManipulatorSettings.serializer(), generateDefinitions = false)
+        val config: SchemaGeneratorConfig = SchemaGeneratorConfigBuilder(SchemaVersion.DRAFT_2019_09, OptionPreset.PLAIN_JSON)
+            .with(JacksonModule())
+            .with(JakartaValidationModule(INCLUDE_PATTERN_EXPRESSIONS, NOT_NULLABLE_FIELD_IS_REQUIRED))
+            .build()
+        val generator = SchemaGenerator(config)
+        val jsonSchema: JsonNode = generator.generateSchema(MegaManipulatorSettings::class.java)
+        val jsonSchemaString = SerializationHolder.readable.writeValueAsString(jsonSchema)
 
         // baseFile.writeText(generatedSchema)
-        assertEquals(fileSystemSchema, generatedSchema)
+        assertThat(fileSystemSchema, equalTo(jsonSchemaString))
     }
 }